+++ priv/repo/migrations/20231016125153_create_irc_tables_table.exs
@@ -4,7 +4,7 @@ defmodule SuperPoker.Repo.Migrations.CreateIrcTablesTable do
   def change do
     create table(:irc_tables, primary_key: false) do
       # 无需id做主键game_id即可
-      add :game_id, :integer, null: false, primary_key: true
+      add :game_id, :bigint, null: false, primary_key: true
       add :blind, :integer, null: false
       add :pot_after_preflop, :integer, null: false
       add :pot_after_flop, :integer, null: false
